requirements.txt
flask>=3.0
pandas>=2.0
python-dotenv>=1.0
fpdf2>=2.7
